move yolo object detection, update pytorch
