+++ Samples/Tutorial Samples/YOLOv4ObjectDetection/.gitattributes
@@ -1,0 +1,1 @@
+*.onnx filter=lfs diff=lfs merge=lfs -text
