+++ Samples/Tutorial Samples/YOLOv4ObjectDetection/YOLOv4ObjectDetection.sln
@@ -1,0 +1,51 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30523.141
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YOLOv4ObjectDetection", "YOLOv4ObjectDetection\YOLOv4ObjectDetection.csproj", "{0084EDC5-F980-4BCE-8762-8259AB99CA17}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|ARM = Debug|ARM
+		Debug|ARM64 = Debug|ARM64
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|ARM = Release|ARM
+		Release|ARM64 = Release|ARM64
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Debug|ARM.ActiveCfg = Debug|ARM
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Debug|ARM.Build.0 = Debug|ARM
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Debug|ARM.Deploy.0 = Debug|ARM
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Debug|ARM64.Build.0 = Debug|ARM64
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Debug|ARM64.Deploy.0 = Debug|ARM64
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Debug|x64.ActiveCfg = Debug|x64
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Debug|x64.Build.0 = Debug|x64
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Debug|x64.Deploy.0 = Debug|x64
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Debug|x86.ActiveCfg = Debug|x86
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Debug|x86.Build.0 = Debug|x86
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Debug|x86.Deploy.0 = Debug|x86
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Release|ARM.ActiveCfg = Release|ARM
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Release|ARM.Build.0 = Release|ARM
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Release|ARM.Deploy.0 = Release|ARM
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Release|ARM64.ActiveCfg = Release|ARM64
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Release|ARM64.Build.0 = Release|ARM64
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Release|ARM64.Deploy.0 = Release|ARM64
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Release|x64.ActiveCfg = Release|x64
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Release|x64.Build.0 = Release|x64
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Release|x64.Deploy.0 = Release|x64
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Release|x86.ActiveCfg = Release|x86
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Release|x86.Build.0 = Release|x86
+		{0084EDC5-F980-4BCE-8762-8259AB99CA17}.Release|x86.Deploy.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {B95D9F41-0A41-4F8D-8074-FF361E90A155}
+	EndGlobalSection
+EndGlobal
