--- Samples/YOLOv4ObjectDetection/.gitattributes
@@ -1,1 +1,0 @@
-*.onnx filter=lfs diff=lfs merge=lfs -text
