--- Samples/YOLOv4ObjectDetection/README.md
@@ -1,27 +1,0 @@
-# YOLOv4 Object Detection Sample
-
-This UWP C# app uses the YOLOv4 model to detect objects in video streams. 
-
-The [YOLOv4 model](https://github.com/hunglc007/tensorflow-yolov4-tflite) is trained with [Tensorflow-DML](https://github.com/microsoft/tensorflow-directml), and converted to ONNX through [tf2onnx](https://github.com/onnx/tensorflow-onnx)
-
-For how-tos, tutorials and additional information, see the [Windows ML documentation](https://docs.microsoft.com/windows/ai/).
-
-## Requirements
-
-- [Visual Studio 2017 - 15.4 or higher](https://developer.microsoft.com/en-us/windows/downloads)
-- [Windows 10 - Build 18362 or higher](https://www.microsoft.com/en-us/software-download/windowsinsiderpreviewiso)
-- [Windows SDK - Build 18362 or higher](https://www.microsoft.com/en-us/software-download/windowsinsiderpreviewSDK)
-
-
-## Contributing
-
-We're always looking for your help to fix bugs and improve the samples. Create a pull request, and we'll be happy to take a look.
-
-This project has adopted the [Microsoft Open Source Code of Conduct](https://opensource.microsoft.com/codeofconduct/).
-For more information see the [Code of Conduct FAQ](https://opensource.microsoft.com/codeofconduct/faq/) or
-contact [opencode@microsoft.com](mailto:opencode@microsoft.com) with any additional questions or comments.
-
-## License
-
-MIT. See [LICENSE file](https://github.com/Microsoft/Windows-Machine-Learning/blob/master/LICENSE).
-
