--- Samples/YOLOv4ObjectDetection/YOLOv4ObjectDetection/App.xaml
@@ -1,7 +1,0 @@
-﻿<Application
-    x:Class="YOLOv4ObjectDetection.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:YOLOv4ObjectDetection">
-
-</Application>
