--- Samples/YOLOv4ObjectDetection/YOLOv4ObjectDetection/MainPage.xaml
@@ -1,18 +1,0 @@
-﻿<Page
-    x:Class="YOLOv4ObjectDetection.MainPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:YOLOv4ObjectDetection"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-
-    <Grid>
-        <Button x:Name="button_go" Content="Go" Margin="78,74,0,0" VerticalAlignment="Top" Click="button_go_Click"/>
-        <StackPanel Orientation="Horizontal">
-            <CaptureElement x:Name="WebCam" Stretch="Fill"  VerticalAlignment="Center" HorizontalAlignment="Left" Width="416" Height="416"/>
-        </StackPanel>
-        <Canvas x:Name="OverlayCanvas" VerticalAlignment="Center" HorizontalAlignment="Left" Width="416" Height="416"/>
-    </Grid>
-</Page>
