--- Samples/YOLOv4ObjectDetection/YOLOv4ObjectDetection/Package.appxmanifest
@@ -1,51 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-
-<Package
-  xmlns="http://schemas.microsoft.com/appx/manifest/foundation/windows10"
-  xmlns:mp="http://schemas.microsoft.com/appx/2014/phone/manifest"
-  xmlns:uap="http://schemas.microsoft.com/appx/manifest/uap/windows10"
-  IgnorableNamespaces="uap mp">
-
-  <Identity
-    Name="633f1357-f236-48e9-91f1-2c99eb682f7f"
-    Publisher="CN=brianma"
-    Version="1.0.0.0" />
-
-  <mp:PhoneIdentity PhoneProductId="633f1357-f236-48e9-91f1-2c99eb682f7f" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
-
-  <Properties>
-    <DisplayName>YOLOv4ObjectDetection</DisplayName>
-    <PublisherDisplayName>brianma</PublisherDisplayName>
-    <Logo>Assets\StoreLogo.png</Logo>
-  </Properties>
-
-  <Dependencies>
-    <TargetDeviceFamily Name="Windows.Universal" MinVersion="10.0.0.0" MaxVersionTested="10.0.0.0" />
-  </Dependencies>
-
-  <Resources>
-    <Resource Language="x-generate"/>
-  </Resources>
-
-  <Applications>
-    <Application Id="App"
-      Executable="$targetnametoken$.exe"
-      EntryPoint="YOLOv4ObjectDetection.App">
-      <uap:VisualElements
-        DisplayName="YOLOv4ObjectDetection"
-        Square150x150Logo="Assets\Square150x150Logo.png"
-        Square44x44Logo="Assets\Square44x44Logo.png"
-        Description="YOLOv4ObjectDetection"
-        BackgroundColor="transparent">
-        <uap:DefaultTile Wide310x150Logo="Assets\Wide310x150Logo.png"/>
-        <uap:SplashScreen Image="Assets\SplashScreen.png" />
-      </uap:VisualElements>
-    </Application>
-  </Applications>
-
-  <Capabilities>
-    <Capability Name="internetClient" />
-    <DeviceCapability Name="webcam"/>
-    <DeviceCapability Name="microphone"/>
-  </Capabilities>
-</Package>
